Покращено структуру проєкта
