+++ game.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-python3 main.py
+python3 code/main.py
